Move User class to login module
